pyproject.toml
[project]
name = "proj_garagem"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "renascm_", email = "renatochagas.m@gmail.com_" }]
dependencies = [
    "django>=4.2.20",
    "djangorestframework>=3.16.0",
    "django-extensions>=4.1",
    "python-dotenv>=1.1.0",
    "netifaces>=0.11.0",
    "dj-database-url>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "uvicorn>=0.34.2",
    "gunicorn>=23.0.0",
    "whitenoise>=6.9.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework-simplejwt>=5.5.0",
    "setuptools>=80.9.0",
    "Pillow>=11.2.1",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
    "drf-spectacular>=0.28.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
    "pygraphviz>=1.11"
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python src/manage.py runserver 0.0.0.0:19003"
createsuperuser = "python src/manage.py createsuperuser"
pre_migrate = "python src/manage.py makemigrations"
migrate = "python src/manage.py migrate"
startapp = "python src/manage.py startapp {args}"
filename = "requirements.txt"
